refactored importpedestrian refactored importpedestrian to open and read from compressed nwp file standarized variable naming conventions and removed all unneeded json paramaters. fixed variable spelling mistake in importnetwork migrated common methods into the common module
